[ALPS04416536] WiFi: Whole chip reset development
1. Refine conninfra bus hang check
2. Add log for trigger assert/ whole chip reset

MTK-Commit-Id: d36245ba0ca3882453fed5f4f37d165eb32804b7

CR-Id: ALPS04416536
Feature: [Module]Wi-Fi driver

+++ include/chips/soc3_0.h
@@ -157,7 +157,7 @@ struct ROM_EMI_HEADER {
 */
 extern struct platform_device *g_prPlatDev;
 #if (CFG_SUPPORT_CONNINFRA == 1)
-extern u_int8_t g_IsConninfraBusHang;
+extern int g_IsConninfraBusHang;
 extern u_int8_t g_IsWfsysBusHang;
 extern struct completion g_triggerComp;
 #endif
